Indices now take index_t size.

+++ src/tensor/tensor_foldin.cc
@@ -36,7 +36,7 @@ void do_foldin_into(Tensor<elt_t> &output, const Tensor<elt_t> &a, int _ndx1,
   const index rankb = b.rank();
   index ndx1 = Dimensions::normalize_index(_ndx1, ranka);
   index ndx2 = Dimensions::normalize_index(_ndx2, rankb);
-  Indices new_dims(static_cast<size_t>(std::max(ranka + rankb - 2, index(1))));
+  Indices new_dims(std::max(ranka + rankb - 2, index(1)));
   /*
      * Since we use row-major order, in which the first
      * index varies faster, we nest the loops beginning with the last index,
